requirements.txt
click==8.1.7
Pillow==10.0.1
